refactor(weather): remove DTO wrappers, use native WeatherKit types directly
